Pull test bootstrap generation out of FlutterPlatform
So that it can be used in 'test/runner.dart'
